Separate PeopleList from PeopleTaable, fix sort
